Version .9.2.0
This release was a major update to previous versions, in that it
migrated away from the use of the PyGreSQL driver and to pg8000. It
also fundamentally changed the runtime architecture so that the utility
can be run as a [scheduled Lambda
function](https://github.com/awslabs/amazon-redshift-utils/tree/master/s
rc/LambdaRunner).

Furthermore, this version significantly reduces the amount of work that
the analyzer will attempt to do. It will only attempt to change
encoding modification for those tables which contain data, as before,
but now will also only attempt to analyze those tables which contain
unencoded columns other than the first column of the sort key. It will
also suppress modifications of any tables where the outcome of the
analysis is the same as the table in place already - sounds obvious but
previous versions would do a migration to an identical table structure