Add support for materialized views in v_generate_view_ddl
Materialized views cannot be dropped using `DROP VIEW` but instead
need to use `DROP MATERIALIZED VIEW` this change will make sure that
the `DROP` in the comment is appropriately changed.

The provided DDL is also fixed; prior to this change the DDL was
preceded with a `CREATE OR REPLACE VIEW <schemaname>.<viewname>`.
This change will generate correct DDL.

To do this we check the generated view definition with exact positions
for the keywords since those keywords could be part of the view
definition itself.

This new version has been tested with the following views:
 - regular view
 - late binding view
 - materialized view