feat: Add view v_view_table_column_dependency
Add a view to get the the views that depend on specific columns.
This is useful when you have columns from which you want to migrate away.
For example after creating new columns with more efficient datatypes.

Views depending on the old column can be easily identified using this
view and thus allows the DBA to rewrite them to benefit the workload
by using more efficient new columns.

Tested on Redshift version 1.0.25813.